Split git command functions out of ginclient package
