+++ ginclient/pkgdoc.go
@@ -1,0 +1,4 @@
+/*
+Package ginclient augments the web package with functions to interact with GIN Gogs (https://github.com/G-Node/gogs) specifically.
+*/
+package ginclient
