+++ gincmd/getcontentcmd.go
@@ -2,6 +2,7 @@ package gincmd
 
 import (
 	ginclient "github.com/G-Node/gin-cli/ginclient"
+	"github.com/G-Node/gin-cli/git"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
@@ -10,13 +11,13 @@ func getContent(cmd *cobra.Command, args []string) {
 	jsonout, _ := cmd.Flags().GetBool("json")
 	gincl := ginclient.New(util.Config.GinHost)
 	requirelogin(cmd, gincl, !jsonout)
-	if !ginclient.IsRepo() {
+	if !git.IsRepo() {
 		util.Die("This command must be run from inside a gin repository.")
 	}
 
 	gincl.GitHost = util.Config.GitHost
 	gincl.GitUser = util.Config.GitUser
-	getcchan := make(chan ginclient.RepoFileStatus)
+	getcchan := make(chan git.RepoFileStatus)
 	go gincl.GetContent(args, getcchan)
 	formatOutput(getcchan, jsonout)
 }
