+++ gincmd/lockcmd.go
@@ -2,17 +2,18 @@ package gincmd
 
 import (
 	ginclient "github.com/G-Node/gin-cli/ginclient"
+	"github.com/G-Node/gin-cli/git"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
 
 func lock(cmd *cobra.Command, args []string) {
 	jsonout, _ := cmd.Flags().GetBool("json")
-	if !ginclient.IsRepo() {
+	if !git.IsRepo() {
 		util.Die("This command must be run from inside a gin repository.")
 	}
 	gincl := ginclient.New(util.Config.GinHost)
-	lockchan := make(chan ginclient.RepoFileStatus)
+	lockchan := make(chan git.RepoFileStatus)
 	go gincl.LockContent(args, lockchan)
 	formatOutput(lockchan, jsonout)
 }
