+++ gincmd/removecontentcmd.go
@@ -2,6 +2,7 @@ package gincmd
 
 import (
 	ginclient "github.com/G-Node/gin-cli/ginclient"
+	"github.com/G-Node/gin-cli/git"
 	"github.com/G-Node/gin-cli/util"
 	"github.com/spf13/cobra"
 )
@@ -10,15 +11,15 @@ func remove(cmd *cobra.Command, args []string) {
 	jsonout, _ := cmd.Flags().GetBool("json")
 	gincl := ginclient.New(util.Config.GinHost)
 	requirelogin(cmd, gincl, !jsonout)
-	if !ginclient.IsRepo() {
+	if !git.IsRepo() {
 		util.Die("This command must be run from inside a gin repository.")
 	}
 	gincl.GitHost = util.Config.GitHost
 	gincl.GitUser = util.Config.GitUser
-	lockchan := make(chan ginclient.RepoFileStatus)
+	lockchan := make(chan git.RepoFileStatus)
 	go gincl.LockContent(args, lockchan)
 	formatOutput(lockchan, jsonout)
-	rmchan := make(chan ginclient.RepoFileStatus)
+	rmchan := make(chan git.RepoFileStatus)
 	go gincl.RemoveContent(args, rmchan)
 	formatOutput(rmchan, jsonout)
 }
